requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
clickhouse-connect==0.8.5
colorama==0.4.6
fastapi==0.115.3
h11==0.14.0
idna==3.10
lz4==4.3.3
mangum==0.19.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
starlette==0.41.0
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
zstandard==0.23.0
